refactored create logic
