+++ src/intentDelegates/addItem.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const createItem = require('./../helperDelegates/createItem');
-
 module.exports = function(handler, table) {
 
     const slots = handler.event.request.intent.slots
@@ -14,7 +12,39 @@ module.exports = function(handler, table) {
                 })
         .then(function(resp) {
             if (resp == undefined) {
-              createItem(handler, table);
+                var record = {
+                    userId: handler.event.session.user.userId,
+                    createTime: Math.floor(Date.now() / 1000)
+                }
+
+                const slots = handler.event.request.intent.slots
+
+                if (slots.Item && slots.Item.value) {
+                    record.itemName = slots.Item.value.toLowerCase();
+
+                    if (slots.Quantity && slots.Quantity.value) {
+                        record.quantity = slots.Quantity.value;
+                    } else {
+                        record.quantity = 1;
+                    }
+                    if (slots.Location && slots.Location.value) {
+                        record.location = slots.Location.value
+                    }
+
+                    table
+                    .insert(record)
+                    .then(function(resp) {
+                        handler.emit('Affirmative');
+                    })
+                    .catch(function(err) {
+                        console.log(err);
+                        handler.emit('Error');
+                    });
+
+                } else {
+                    console.log("error with itemName slot");
+                    handler.emit('Error')
+                }
             } else {
               if (slots.Quantity && slots.Quantity.value) {
                 setQuantity = slots.Quantity.value;
